// src/api/schema.d.ts
export interface components {
    schemas: {
        Post: {
            id: number;
            userId: number;
            title: string;
            body: string;
        };
        CreatePostDto: {
            userId: number;
            title: string;
            body: string;
        };
        UpdatePostDto: {
            userId?: number;
            title?: string;
            body?: string;
        };
        Comment: {
            id: number;
            postId: number;
            name: string;
            /** Format: email */
            email: string;
            body: string;
        };
        CreateCommentDto: {
            postId: number;
            name: string;
            /** Format: email */
            email: string;
            body: string;
        };
        UpdateCommentDto: {
            postId?: number;
            name?: string;
            /** Format: email */
            email?: string;
            body?: string;
        };
        User: {
            id: number;
            name: string;
            username: string;
            /** Format: email */
            email: string;
            address: {
                street: string;
                suite: string;
                city: string;
                zipcode: string;
                geo: {
                    lat: string;
                    lng: string;
                };
            };
            phone: string;
            website: string;
            company: {
                name: string;
                catchPhrase: string;
                bs: string;
            };
        };
        CreateUserDto: {
            name: string;
            username: string;
            /** Format: email */
            email: string;
            address: {
                street: string;
                suite: string;
                city: string;
                zipcode: string;
                geo: {
                    lat: string;
                    lng: string;
                };
            };
            phone: string;
            website: string;
            company: {
                name: string;
                catchPhrase: string;
                bs: string;
            };
        };
        UpdateUserDto: {
            name?: string;
            username?: string;
            /** Format: email */
            email?: string;
            address?: {
                street: string;
                suite: string;
                city: string;
                zipcode: string;
                geo: {
                    lat: string;
                    lng: string;
                };
            };
            phone?: string;
            website?: string;
            company?: {
                name: string;
                catchPhrase: string;
                bs: string;
            };
        };
        Todo: {
            id: number;
            userId: number;
            title: string;
            completed: boolean;
        };
        Album: {
            id: number;
            userId: number;
            title: string;
        };
        CreateTodoDto: {
            userId: number;
            title: string;
            completed: boolean;
        };
        UpdateTodoDto: {
            userId?: number;
            title?: string;
            completed?: boolean;
        };
        CreateAlbumDto: {
            userId: number;
            title: string;
        };
        UpdateAlbumDto: {
            userId?: number;
            title?: string;
        };
        Photo: {
            id: number;
            albumId: number;
            title: string;
            /** Format: uri */
            url: string;
            /** Format: uri */
            thumbnailUrl: string;
        };
        CreatePhotoDto: {
            albumId: number;
            title: string;
            /** Format: uri */
            url: string;
            /** Format: uri */
            thumbnailUrl: string;
        };
        UpdatePhotoDto: {
            albumId?: number;
            title?: string;
            /** Format: uri */
            url?: string;
            /** Format: uri */
            thumbnailUrl?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}